start app, shoes4, and jruby
Messed around with ruby shoes a bit. Shoes3 is a
native app that has a packager. Shoes4 uses jruby
and Java libs for GUI. Shoes4 is where it is at.

And btw switching to JRuby improved the time to
calculate arrangements for a hand

MRI:   0.072 ms
JRuby: 0.039 ms

So that is pretty nice.

Found open source Mah Jong tile images, so a big thank you to
FluffyStuff: https://github.com/FluffyStuff/riichi-mahjong-tiles